Repo Initialization
This is a copy of a repo that exists here: https://github.com/lucasejones/Web-Development-Practice/tree/main/freeCodeCamp%20projects/Star-Wars-Planet-Explorer

I made this one because it was a subdirectory in the other repo, and I wanted to facilitate easier use without someone needing to clone 40 other projects. However, the version history was not transferred over, so the link above will show that up to this point.