Functioning List Items
I made PlanetBox a child of PlanetsList and now the links correctly display the planet and corresponding info. Additionally, the visit random planet functionality is still intact.

However, the visual indicator I implemented to show what the currently selected planet is only works if not using the visit random planet button. to solve this I will likely need to make a common parent component (my original plan) instead of having PlanetBox be a child of PlanetsList.