State Refactor
I created a PlanetsData component to contain the two components of PlanetBox and PlanetsList. I did this in order for them to both respond to changes to the state of what the current planet was, as well as whether the random button was the cause of the current planet.

I also cleaned up some old commented code and made the syntax more adherent to the codebase's general stylings.