Re-introduced Key Attribute
I noticed while using the react dev tools profiler that the PlanetsList was taking an oddly long time to render, and on investigation found that I had mistakenly deleted the key attribute in the map function in that component. Restoring this cut the render time to 1/3. (3ms down to 1ms).